requirements.txt
flask==2.3.3
google-api-python-client==2.97.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
pandas==2.0.3
